# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Open Banking API
  version: 1.0.0
  summary: Open Banking-style contract for account information, consent, and payment initiation.
  description: |
    This OpenAPI document defines a small Open Banking-style service. It demonstrates
    account information endpoints, balance and transaction access, consent lifecycle
    operations, payment initiation, reusable component schemas, standardized errors,
    and cursor-based pagination.
servers:
  - url: https://api.example.com/open-banking/v1
    description: Production
  - url: https://sandbox-api.example.com/open-banking/v1
    description: Sandbox

tags:
  - name: System
  - name: Accounts
  - name: Consents
  - name: Payments

security:
  - ApiKeyAuth: []

paths:
  /health:
    get:
      security: []
      tags: [System]
      summary: Health check
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

  /accounts:
    get:
      tags: [Accounts]
      summary: List accounts
      operationId: listAccounts
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: A paginated list of accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /accounts/{accountId}:
    parameters:
      - $ref: '#/components/parameters/AccountIdPathParam'
    get:
      tags: [Accounts]
      summary: Get account by id
      operationId: getAccountById
      responses:
        '200':
          description: Account found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /accounts/{accountId}/balances:
    parameters:
      - $ref: '#/components/parameters/AccountIdPathParam'
    get:
      tags: [Accounts]
      summary: Get account balance
      operationId: getAccountBalance
      responses:
        '200':
          description: Account balance found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /accounts/{accountId}/transactions:
    parameters:
      - $ref: '#/components/parameters/AccountIdPathParam'
    get:
      tags: [Accounts]
      summary: List account transactions
      operationId: listAccountTransactions
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: A paginated list of transactions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /consents:
    post:
      tags: [Consents]
      summary: Create account access consent
      operationId: createConsent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConsentRequest'
      responses:
        '201':
          description: Consent created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /consents/{consentId}:
    parameters:
      - $ref: '#/components/parameters/ConsentIdPathParam'
    get:
      tags: [Consents]
      summary: Get consent by id
      operationId: getConsentById
      responses:
        '200':
          description: Consent found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Consents]
      summary: Revoke consent
      operationId: revokeConsent
      responses:
        '200':
          description: Consent revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /payments:
    post:
      tags: [Payments]
      summary: Initiate payment
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment instruction created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentInstruction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /payments/{paymentId}:
    parameters:
      - $ref: '#/components/parameters/PaymentIdPathParam'
    get:
      tags: [Payments]
      summary: Get payment by id
      operationId: getPaymentById
      responses:
        '200':
          description: Payment found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentInstruction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Demo API key required for Open Banking resource endpoints. The same key may also be supplied as an Authorization Bearer token by the running sample API.
  parameters:
    AccountIdPathParam:
      in: path
      name: accountId
      required: true
      schema:
        type: string
        format: uuid
    ConsentIdPathParam:
      in: path
      name: consentId
      required: true
      schema:
        type: string
        format: uuid
    PaymentIdPathParam:
      in: path
      name: paymentId
      required: true
      schema:
        type: string
        format: uuid
    CursorParam:
      in: query
      name: cursor
      required: false
      schema:
        type: string
    PageSizeParam:
      in: query
      name: pageSize
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  responses:
    BadRequest:
      description: Invalid request payload or query parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: API key is missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: API key is invalid or not authorized.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    HealthStatus:
      type: object
      required: [status, timestamp]
      properties:
        status:
          type: string
          enum: [ok]
        timestamp:
          type: string
          format: date-time
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: decimal
        currency:
          type: string
          minLength: 3
          maxLength: 3
    Account:
      type: object
      required: [id, bankId, accountNumberMasked, displayName, type, status, currency, createdAt]
      properties:
        id:
          type: string
          format: uuid
        bankId:
          type: string
        accountNumberMasked:
          type: string
        displayName:
          type: string
        type:
          type: string
          enum: [Current, Savings, CreditCard]
        status:
          type: string
          enum: [Enabled, Disabled]
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
    Balance:
      type: object
      required: [accountId, available, current, asOf]
      properties:
        accountId:
          type: string
          format: uuid
        available:
          $ref: '#/components/schemas/Money'
        current:
          $ref: '#/components/schemas/Money'
        asOf:
          type: string
          format: date-time
    Transaction:
      type: object
      required: [id, accountId, bookedAt, description, merchantName, amount, category, reference]
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        bookedAt:
          type: string
          format: date-time
        description:
          type: string
        merchantName:
          type: string
        amount:
          $ref: '#/components/schemas/Money'
        category:
          type: string
        reference:
          type: string
    Consent:
      type: object
      required: [id, customerId, permissions, status, expiresAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [AwaitingAuthorization, Authorized, Revoked, Expired]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    PaymentInstruction:
      type: object
      required: [id, consentId, debtorAccountId, creditorAccountNumber, creditorName, amount, remittanceInformation, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        consentId:
          type: string
          format: uuid
        debtorAccountId:
          type: string
        creditorAccountNumber:
          type: string
        creditorName:
          type: string
        amount:
          $ref: '#/components/schemas/Money'
        remittanceInformation:
          type: string
        status:
          type: string
          enum: [Pending, Accepted, Rejected]
        createdAt:
          type: string
          format: date-time
    CreateConsentRequest:
      type: object
      required: [customerId, permissions, expiresAt]
      properties:
        customerId:
          type: string
        permissions:
          type: array
          minItems: 1
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
      additionalProperties: false
    CreatePaymentRequest:
      type: object
      required: [consentId, debtorAccountId, creditorAccountNumber, creditorName, amount, remittanceInformation]
      properties:
        consentId:
          type: string
          format: uuid
        debtorAccountId:
          type: string
        creditorAccountNumber:
          type: string
        creditorName:
          type: string
        amount:
          $ref: '#/components/schemas/Money'
        remittanceInformation:
          type: string
      additionalProperties: false
    AccountListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        nextCursor:
          type: string
    TransactionListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        nextCursor:
          type: string
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required: [field, issue]
      properties:
        field:
          type: string
        issue:
          type: string